% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DuoFatherOffspringMendelianErrorCorrection.R
\name{DuoFatherOffspringMendelError}
\alias{DuoFatherOffspringMendelError}
\title{Step 1 of 3 : Mendel Error detection in father-offspring Duos and set it to missing}
\usage{
DuoFatherOffspringMendelError(vcf, ped, ncore)
}
\arguments{
\item{vcf}{Phased genotype vcf file for mother-offspring duos with full path [mandatory]}

\item{ped}{Tab delimited Pedigree file with full path having four columns (FamilyID,IndividualID,FatherID,MotherID) with NO header and 0 missing for missing family member [mandatory]}

\item{n}{Number of cores to use with parallel processing (default 4) [optional]}
}
\description{
Scan each father-offspring duo for Mendel errors and set such sites to missing in all family members using parallel computing
Takes in phased vcf files of duo family data, pedigree file and number of cores (optional, default is 4 cores)
}
\note{
Automatically creates dir named 'MendelErrorFixed/FatherOffsprings' one level above the data dir for the result file of this process
}
\examples{
DuoFatherOffspringMendelError("/my/dir/data/myphased.vcf","/my/dir/data/myped.txt",2)
DuoFatherOffspringMendelError("/my/dir/data/myphased.vcf","/my/dir/data/myped.txt")
}
\author{
Bharati Jadhav
}

