% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DuoFatherOffspringPhaseCorrect.R
\name{DuoFatherOffspringPhaseCorrect}
\alias{DuoFatherOffspringPhaseCorrect}
\title{Step 2 of 3: Sliding window algorithm to detect and fix phase switch error in father-offspring duos}
\usage{
DuoFatherOffspringPhaseCorrect(vcf, ped, ncore)
}
\arguments{
\item{vcf}{Mendel error fixed Phased genotype vcf file for Duo family with full path [mandatory]}

\item{ped}{Tab delimited Pedigree file with full path having four columns (FamilyID,IndividualID,FatherID,MotherID) with NO header and 0 missing for missing family member [mandatory]}

\item{n}{Number of cores to use with parallel processing (default 4) [optional]}
}
\description{
Scan each father-offpsring duo for to detect and fix phase switch error in offspring using sliding window (size 100 and step 50 sites) using parallel computing
Takes in phased vcf files of trio familyd data, pedigree file and number of cores (optional, default is 4 cores)
}
\note{
Automatically creates dir named 'PhaseSwitchSlidingWindow/FatherOffsprings' one level above the data dir for the result file of this process
}
\examples{
DuoFatherOffspringPhaseCorrect("/my/dir/MendelErrorFixed/FatherOffsprings/myphased.MendelErrorFixed.vcf","/my/dir/data/myped.txt",2)
DuoFatherOffspringPhaseCorrect("/my/dir/MendelErrorFixed/FatherOffsprings/myphased.MendelErrorFixed.vcf","/my/dir/data/myped.txt")
}
\author{
Bharati Jadhav
}

